Reworked Paths to use static string, made little sense to use StringIds as we want paths to capture the actual underlying string Added base singleton classes. Added skeleton Module code - modules give us coarse grained entry points to major engine systems. Cleaned up warnings and compile errors, untracked redundant files and vendor folders

+++ Velocity/src/Engine/Renderer/Mesh/Model.h
@@ -35,7 +35,7 @@ namespace Vct
         Mesh* GetMesh(uint8_t meshId) const;
         MaterialInstance* GetMaterial(uint8_t materialId) const;
 
-        FORCEINLINE int GetNumMeshes() const { return m_Meshes.size(); }
+        FORCEINLINE int GetNumMeshes() const { return static_cast<int>(m_Meshes.size()); }
         RES_TYPE(ResType_Model);
     protected:
         /* Resource Interface */
